requirements.txt
pytest==8.3.4
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
setuptools==41.2.0
